Revert "Refactor files to src folder"
This reverts commit 20c3c5353e294d36037f871b77a1170b913dac01.
